+++ Cargo.toml
@@ -21,7 +21,8 @@ tracing = "0.1"
 
 [dependencies.warp]
 version = "0.3"
-features = ["tls"]
+features = ["tls", "multipart"]
+default-features = false
 
 [dependencies.assembly-core]
 version = "0.2.1"
